+++ src/algorithm/array/slidingWindow/checkInclusion.ts
@@ -1,38 +1,38 @@
 // https://leetcode-cn.com/problems/permutation-in-string/
 // 567
 export function checkInclusion(t: string, s: string): boolean {
-  const need = new Map<string, number>();
-  const slidingWindow = new Map<string, number>();
+  const needMap = new Map<string, number>();
+  const countMap = new Map<string, number>();
   let left = 0;
   let right = 0;
   let windowSize = 0;
 
   for (let i = 0; i < t.length; i++) {
-    need.set(t[i], (need.get(t[i]) || 0) + 1);
+    needMap.set(t[i], (needMap.get(t[i]) || 0) + 1);
   }
 
   while (right < s.length) {
     const rightChar = s[right];
     right++;
-    if (need.has(rightChar)) {
-      slidingWindow.set(rightChar, (slidingWindow.get(rightChar) || 0) + 1);
-      if (slidingWindow.get(rightChar) === need.get(rightChar)) {
+    if (needMap.has(rightChar)) {
+      countMap.set(rightChar, (countMap.get(rightChar) || 0) + 1);
+      if (countMap.get(rightChar) === needMap.get(rightChar)) {
         windowSize++;
       }
     }
 
-    while (right - left >= t.length) {
-      if (windowSize === need.size) {
+    while (right - left === t.length) {
+      if (windowSize === needMap.size) {
         return true;
       }
 
       const leftChar = s[left];
       left++;
-      if (need.has(leftChar)) {
-        if (slidingWindow.get(leftChar) === need.get(leftChar)) {
+      if (needMap.has(leftChar)) {
+        if (countMap.get(leftChar) === needMap.get(leftChar)) {
           windowSize--;
         }
-        slidingWindow.set(leftChar, slidingWindow.get(leftChar) - 1);
+        countMap.set(leftChar, countMap.get(leftChar) - 1);
       }
     }
   }