+++ src/algorithm/array/slidingWindow/lengthOfLongestSubstring.ts
@@ -1,7 +1,7 @@
 // https://leetcode-cn.com/problems/longest-substring-without-repeating-characters/submissions/
 // 3
 export function lengthOfLongestSubstring(s: string): number {
-  const slidingWindow = new Map<string, number>();
+  const countMap = new Map<string, number>();
   let left = 0;
   let right = 0;
   let length = 0;
@@ -9,12 +9,12 @@ export function lengthOfLongestSubstring(s: string): number {
   while (right < s.length) {
     const rightChar = s[right];
     right++;
-    slidingWindow.set(rightChar, (slidingWindow.get(rightChar) || 0) + 1);
+    countMap.set(rightChar, (countMap.get(rightChar) || 0) + 1);
 
-    while (slidingWindow.get(rightChar) > 1) {
+    while (countMap.get(rightChar) > 1) {
       const leftChar = s[left];
       left++;
-      slidingWindow.set(leftChar, slidingWindow.get(leftChar) - 1);
+      countMap.set(leftChar, countMap.get(leftChar) - 1);
     }
 
     length = Math.max(length, right - left);